requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
pillow==8.4.0
python-docx==0.8.11
reportlab==3.6.1
pyheif==0.6.0
python-multipart==0.0.5
pypandoc==1.13
pydub==0.25.1
aiofiles==23.2.1
pandas==2.2.2
